Improve settings layout
