+++ .idea/assetWizardSettings.xml
@@ -80,8 +80,9 @@
                         </option>
                         <option name="values">
                           <map>
-                            <entry key="outputName" value="ic_wifi" />
-                            <entry key="sourceFile" value="$USER_HOME$/Downloads/pin-outline.svg" />
+                            <entry key="assetSourceType" value="FILE" />
+                            <entry key="outputName" value="ic_file_alert" />
+                            <entry key="sourceFile" value="$USER_HOME$/Downloads/file-alert.svg" />
                           </map>
                         </option>
                       </PersistentState>
