+++ app/src/main/AndroidManifest.xml
@@ -82,7 +82,7 @@
             android:parentActivityName=".AboutActivity"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".Settings.SettingsActivity"
             android:label="@string/title_activity_settings"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
