+++ app/src/main/java/ca/pkay/rcloneexplorer/Settings/SettingsActivity.java
@@ -1,0 +1,166 @@
+package ca.pkay.rcloneexplorer.Settings;
+
+import android.app.ActivityManager;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.Bundle;
+import android.preference.PreferenceManager;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentTransaction;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+
+import ca.pkay.rcloneexplorer.CustomColorHelper;
+import ca.pkay.rcloneexplorer.R;
+
+public class SettingsActivity extends AppCompatActivity implements  SettingsFragment.OnSettingCategorySelectedListener,
+                                                                    LookAndFeelSettingsFragment.OnThemeHasChanged {
+
+    public final static String THEME_CHANGED = "ca.pkay.rcexplorer.SettingsActivity.THEME_CHANGED";
+    private final String SAVED_THEME_CHANGE = "ca.pkay.rcexplorer.SettingsActivity.OUTSTATE_THEME_CHANGED";
+    private final String SAVED_FRAGMENT = "ca.pkay.rcexplorer.SettingsActivity.RESTORE_FRAGMENT";
+    private boolean themeHasChanged;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        applyTheme();
+        setContentView(R.layout.activity_settings);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayHomeAsUpEnabled(true);
+            actionBar.setDisplayShowHomeEnabled(true);
+        }
+
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        for (int i = 0; i < fragmentManager.getBackStackEntryCount(); i++) {
+            fragmentManager.popBackStack();
+        }
+
+        startSettingsFragment();
+
+        if (savedInstanceState != null) {
+            Fragment fragment = getSupportFragmentManager().findFragmentByTag(SAVED_FRAGMENT);
+            if (fragment != null) {
+                restoreFragment(fragment);
+            }
+        }
+
+        themeHasChanged = savedInstanceState != null && savedInstanceState.getBoolean(SAVED_THEME_CHANGE, false);
+        Intent returnData = new Intent();
+        returnData.putExtra(THEME_CHANGED, themeHasChanged);
+        setResult(RESULT_OK, returnData);
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        outState.putBoolean(SAVED_THEME_CHANGE, themeHasChanged);
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+    }
+
+    private void applyTheme() {
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        int customPrimaryColor = sharedPreferences.getInt(getString(R.string.pref_key_color_primary), -1);
+        int customAccentColor = sharedPreferences.getInt(getString(R.string.pref_key_color_accent), -1);
+        boolean isDarkTheme = sharedPreferences.getBoolean(getString(R.string.pref_key_dark_theme), false);
+        getTheme().applyStyle(CustomColorHelper.getPrimaryColorTheme(this, customPrimaryColor), true);
+        getTheme().applyStyle(CustomColorHelper.getAccentColorTheme(this, customAccentColor), true);
+        if (isDarkTheme) {
+            getTheme().applyStyle(R.style.DarkTheme, true);
+        } else {
+            getTheme().applyStyle(R.style.LightTheme, true);
+        }
+
+        // set recents app color to the primary color
+        Bitmap bm = BitmapFactory.decodeResource(getResources(), R.mipmap.ic_launcher_round);
+        ActivityManager.TaskDescription taskDesc = new ActivityManager.TaskDescription(getString(R.string.app_name), bm, customPrimaryColor);
+        setTaskDescription(taskDesc);
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        onBackPressed();
+        return true;
+    }
+
+    private void restoreFragment(Fragment fragment) {
+        if (fragment instanceof GeneralSettingsFragment) {
+            startGeneralSettingsFragment();
+        } else if (fragment instanceof LookAndFeelSettingsFragment) {
+            startLookAndFeelSettingsFragment();
+        } else if (fragment instanceof NotificationsSettingsFragment) {
+            startNotificationSettingsFragment();
+        } else if (fragment instanceof LoggingSettingsFragment) {
+            startLoggingSettingsActivity();
+        }
+    }
+
+    private void startSettingsFragment() {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.flFragment, SettingsFragment.newInstance());
+        transaction.commit();
+    }
+
+    private void startGeneralSettingsFragment() {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.flFragment, GeneralSettingsFragment.newInstance(), SAVED_FRAGMENT);
+        transaction.addToBackStack(null);
+        transaction.commit();
+    }
+
+    private void startLookAndFeelSettingsFragment() {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.flFragment, LookAndFeelSettingsFragment.newInstance(), SAVED_FRAGMENT);
+        transaction.addToBackStack(null);
+        transaction.commit();
+    }
+
+    private void startNotificationSettingsFragment() {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.flFragment, NotificationsSettingsFragment.newInstance(), SAVED_FRAGMENT);
+        transaction.addToBackStack(null);
+        transaction.commit();
+    }
+
+    private void startLoggingSettingsActivity() {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.flFragment, LoggingSettingsFragment.newInstance(), SAVED_FRAGMENT);
+        transaction.addToBackStack(null);
+        transaction.commit();
+    }
+
+    @Override
+    public void onSettingCategoryClicked(int category) {
+        switch (category) {
+            case SettingsFragment.GENERAL_SETTINGS:
+                startGeneralSettingsFragment();
+                break;
+            case SettingsFragment.LOOK_AND_FEEL_SETTINGS:
+                startLookAndFeelSettingsFragment();
+                break;
+            case SettingsFragment.LOGGING_SETTINGS:
+                startLoggingSettingsActivity();
+                break;
+            case SettingsFragment.NOTIFICATION_SETTINGS:
+                startNotificationSettingsFragment();
+                break;
+        }
+    }
+
+    @Override
+    public void onThemeChanged() {
+        themeHasChanged = true;
+        recreate();
+    }
+}
