--- app/src/main/res/drawable/ic_android_head.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M8,11.5A1.25,1.25 0,0 0,6.75 12.75A1.25,1.25 0,0 0,8 14A1.25,1.25 0,0 0,9.25 12.75A1.25,1.25 0,0 0,8 11.5M16,11.5A1.25,1.25 0,0 0,14.75 12.75A1.25,1.25 0,0 0,16 14A1.25,1.25 0,0 0,17.25 12.75A1.25,1.25 0,0 0,16 11.5M12,7C13.5,7 14.9,7.33 16.18,7.91L18.34,5.75C18.73,5.36 19.36,5.36 19.75,5.75C20.14,6.14 20.14,6.77 19.75,7.16L17.95,8.96C20.41,10.79 22,13.71 22,17H2C2,13.71 3.59,10.79 6.05,8.96L4.25,7.16C3.86,6.77 3.86,6.14 4.25,5.75C4.64,5.36 5.27,5.36 5.66,5.75L7.82,7.91C9.1,7.33 10.5,7 12,7Z"/>
-</vector>
